feat: add optional stop before function to replayq:pop/2 options
Two new optional options has been added to the replayq:pop/2 function:
* stop_before and,
* stop_before_input_accumulator.

The stop_before option is a function where the first argument is the
current item, and the second argument is an accumulator (starts with the
value of stop_before_input_accumulator if specified or #{} if
stop_before_input_accumulator is not specified). The stop_before
function should return true if the pop call should stop adding items
before the current item. If the stop_before function does not return
true, the returned value will be the accumulator argument to the next
call of the stop_before function.